0.64 add geometry shaders

+++ opengl_renderer/data/scenes/2.3_importmodel_testscene.cpp
@@ -70,7 +70,7 @@ void ImportModelTestScene::update() {
 	// move lights radial
 // --------------------------------------------------------------------------------------
 	float distance_multiplier = 3.0f;
-	const int pi = 3.141592f;
+	const float pi = 3.141592f;
 	for (int ii = 0; ii < point_lights.size(); ii++)
 	{
 		// change light position
