+++ opengl_renderer/data/scenes/3.2_blending_testscene.cpp
@@ -118,7 +118,7 @@ void BlendingTestScene::update() {
 	// move lights radial
 	// --------------------------------------------------------------------------------------
 	float distance_multiplier = 3.0f;
-	const int pi = 3.141592f;
+	const float pi = 3.141592f;
 	for (int ii = 0; ii < point_lights.size(); ii++)
 	{
 		// change light position
